[CLN] Refactor code to extract common code between obj and glb
