+++ tests/test_io/test_meshes/test_glf.py
@@ -4,6 +4,7 @@ import pytest
 import dotenv
 
 from subsurface import optional_requirements
+from subsurface.modules.reader.mesh.glb_reader import load_glb_with_trimesh
 from subsurface.modules.reader.mesh.obj_reader import load_obj_with_trimesh
 from tests.conftest import RequirementsLevel
 
@@ -11,7 +12,7 @@ import subsurface
 from subsurface.modules.visualization import to_pyvista_mesh, pv_plot
 
 from subsurface import optional_requirements, TriSurf
-from subsurface.modules.reader.mesh.obj_reader import load_obj_with_trimesh, trimesh_obj_to_unstruct
+from subsurface.modules.reader.mesh.obj_reader import load_obj_with_trimesh, trimesh_to_unstruct
 
 dotenv.load_dotenv()
 
@@ -36,10 +37,9 @@ def test_trimesh_read_glb_complex():
 
     # Trimesh can load GLB/GLTF natively
 
-    trimesh_obj = load_obj_with_trimesh(glb_path, plot=False)
-    ts: subsurface.TriSurf = trimesh_obj_to_unstruct(trimesh_obj)
+    ts: subsurface.TriSurf = load_glb_with_trimesh(glb_path)
     s = to_pyvista_mesh(ts)
-    pv_plot([s], image_2d=False)
+    pv_plot([s], image_2d=True)
     return
 
 
@@ -55,8 +55,7 @@ def test_trimesh_read_glb():
 
     # Trimesh can load GLB/GLTF natively
 
-    trimesh_obj = load_obj_with_trimesh(glb_path, plot=False)
-    ts: subsurface.TriSurf = trimesh_obj_to_unstruct(trimesh_obj)
+    ts = load_glb_with_trimesh(glb_path, plot=False)
     s = to_pyvista_mesh(ts)
     pv_plot([s], image_2d=True)
     return
